requirements.txt
aiohttp
aiohttp-cors
clifford
faiss-cpu
google-generativeai
matplotlib
networkx
numpy
ollama
openai
rich
scikit-learn
scipy
torch
websockets
PyYAML
typing-extensions
fastapi
uvicorn[standard]
